Remove acceswallet container
