--- src/layouts/AccessWalletLayout/containers/AccessMyWalletContainer/index.js
@@ -1,1 +1,0 @@
-export { default } from './AccessMyWalletContainer';
